backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
openai==1.30.1
anthropic==0.25.0
python-dotenv==1.0.1
# Optional: for local Whisper (no API key needed)
# openai-whisper==20231117
# torch
